feedback stuff

+++ Bermuda/Bermuda/WanderAround.cpp
@@ -36,8 +36,7 @@ void WanderAround::execute(Entity* entity, double dt)
 
 	if (movableEntity->getTimeSinceLastAction() < timeWait)
 	{
-		// TODO: fix double to long warning
-		movableEntity->setTimeSinceLastAction(movableEntity->getTimeSinceLastAction() + GameTimer::Instance()->getFrameTime());
+		movableEntity->setTimeSinceLastAction(movableEntity->getTimeSinceLastAction() + static_cast<long>( GameTimer::Instance()->getFrameTime() ));
 	}
 	else {
 		movableEntity->setTimeSinceLastAction(0);
